chore: split loader.ts into files
