+++ packages/icons/src/loaders/Loadable.ts
@@ -1,0 +1,4 @@
+export interface Loadable {
+  fetch(): Promise<string>
+  isLoading(): boolean
+}
